Refactor rooms #97
