+++ src/rooms/room.routes.ts
@@ -1,14 +1,14 @@
-import { Request, Response } from 'express'
+import { Request, Response, Router } from 'express'
 
-import { getBusyRooms, getEventsForRoom } from '../services/room.service'
+import { getBusyRooms, getEventsForRoom } from './room.service'
 import { ROOMS } from '../util/constants'
 
-export const index = async (_req: Request, res: Response) => {
+const index = async (_req: Request, res: Response) => {
   const busyRooms = await getBusyRooms()
   res.render('room/index', { busyRooms, ROOMS })
 }
 
-export const show = async (req: Request, res: Response) => {
+const show = async (req: Request, res: Response) => {
   if (+req.params.id <= 18 && +req.params.id >= 3) {
     res.render('room/calendar', { room: req.params.id })
   } else {
@@ -16,7 +16,7 @@ export const show = async (req: Request, res: Response) => {
   }
 }
 
-export const getGroupsForRoom = async (req: Request, res: Response) => {
+const getGroupsForRoom = async (req: Request, res: Response) => {
   const events = await getEventsForRoom(+req.params.id)
   res.json(
     events.map(event => ({
@@ -27,3 +27,11 @@ export const getGroupsForRoom = async (req: Request, res: Response) => {
     }))
   )
 }
+
+const router = Router()
+
+router.get('/', index)
+router.get('/:id', show)
+router.get('/:id/events', getGroupsForRoom)
+
+export default router
