+++ src/rooms/room.service.ts
